// src/app/client/model/product.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Group } from './group';
import { ImageDesc } from './imageDesc';
import { ProductProperties } from './productProperties';
import { ProductVariation } from './productVariation';


export class Product {
    id?: string;
    groupId?: string;
    group?: Group;
    title?: string;
    description?: string;
    longDescription?: string;
    created?: number;
    lastChange?: number;
    properties?: ProductProperties;
    images?: Array<ImageDesc>;
    isArchived?: boolean;
    variations?: Array<ProductVariation>;
    order?: number;
}