// src/app/client/model/variationParameter.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VariationParameterValue } from './variationParameterValue';


export class VariationParameter {
    id?: string;
    title?: string;
    values?: Array<VariationParameterValue> = new Array<VariationParameterValue>();
}